backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
ecdsa==0.19.2
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
